gateway-api/requirements.txt
# FastAPI e dependências
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0  # Para BaseSettings
email-validator>=2.0.0

# Cliente HTTP para comunicação entre serviços
httpx>=0.24.0

# Autenticação e segurança
python-jose>=3.3.0  # Para JWT
passlib>=1.7.4  # Para hash de senhas
bcrypt>=4.0.1  # Para hash de senhas

# Middleware e utilitários
starlette>=0.27.0
python-multipart>=0.0.6

# Logging e monitoramento
structlog>=23.1.0

# Validação de dados
validators>=0.20.0

# Ferramentas de desenvolvimento
pytest>=7.3.1
pytest-asyncio>=0.21.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0